third_party/blink/web_tests/serial/serialPort_writable.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/mojo/public/mojom/base/unguessable_token.mojom.js"></script>
<script src="file:///gen/third_party/blink/public/mojom/serial/serial.mojom.js"></script>
<script src="resources/serial-test-utils.js"></script>
<script>

serial_test(async (t, fake) => {
  const { port, fakePort } = await getFakeSerialPort(fake);

  assert_equals(port.writable, null);

  await port.open({ baudrate: 9600 });
  assert_true(port.writable instanceof WritableStream);

  port.close();
  assert_equals(port.writable, null);
}, 'open() and close() set and unset SerialPort.writable');

serial_test(async (t, fake) => {
  const { port, fakePort } = await getFakeSerialPort(fake);

  await port.open({ baudrate: 9600 });
  assert_true(port.writable instanceof WritableStream);

  const writer = port.writable.getWriter();
  port.close();

  const data = new Uint8Array([1, 2, 3, 4, 5, 6, 7, 8]);
  await promise_rejects(t, 'InvalidStateError', writer.write(data));
  assert_equals(port.writable, null);
}, 'SerialPort.writable reports errors after close()');

serial_test(async (t, fake) => {
  const { port, fakePort } = await getFakeSerialPort(fake);

  await port.open({ baudrate: 9600 });
  assert_true(port.writable instanceof WritableStream);

  const writer = port.writable.getWriter();
  const data = new Uint8Array([1, 2, 3, 4, 5, 6, 7, 8]);
  let writePromise = writer.write(data);
  let { value, done } = await fakePort.read();
  await writePromise;
  compareArrays(value, data);

  port.close();
  assert_equals(port.writable, null);
}, 'Can write a small amount of data');

serial_test(async (t, fake) => {
  const { port, fakePort } = await getFakeSerialPort(fake);
  // Select a buffer size smaller than the amount of data transferred.
  await port.open({ baudrate: 9600, buffersize: 64 });

  const writer = port.writable.getWriter();
  const data = new Uint8Array(1024);  // Much larger than buffersize above.
  for (let i = 0; i < data.byteLength; ++i)
    data[i] = i & 0xff;
  writer.write(data);
  writer.close();

  const reader = fakePort.readable_.getReader();
  const value = await readAll(reader);
  reader.releaseLock();
  compareArrays(data, value);
}, 'Can read a large amount of data');

</script>
